backend/requirement.txt
fastapi
uvicorn
pydantic
google-generativeai
python-dotenv
pymupdf
python-multipart
chromadb 
sentence-transformers